+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 * Fixed a bug in printing truncated results.
 * Now characters are internally represented as `Word8`s instead of `Char`s.
 * Added a REPL command `\Arity` to print the arity of a function.
-* New built-in functions: `\minMax`, `\pad`, `\slices`.
+* New built-in functions: `\minValue`, `\minMax`, `\maxValue`, `\ordering`, `\pad`, `\slices`.
 
 ### Breaking changes
 
 * `\apply2` is renamed to `\onBoth`.
+* `\interval` can now take two chars as arguments.
 
 ### Breaking changes